Add files via upload
Heart disease is the leading cause of death in the United States. Researchers are using several data mining techniques to help health care professionals in the diagnosis of heart disease. In this project, we will examine the relationship between the maximum heart rate one can achieve during exercise and the likelihood of developing heart disease. Using multiple logistic regression, we will handle the confounding effects of age and gender.

This project uses the Cleveland heart disease dataset.